requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
modbus-tk==1.1.2
django_cleanup